feat-02: add menu page component
