--- src/components/content/Content.css
@@ -1,39 +1,0 @@
-.item-list {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    font-family: Arial, sans-serif;
-    background-color: #f9f9f9;
-    border-radius: 8px;
-}
-
-h1 {
-    text-align: center;
-    color: #333;
-}
-
-.item {
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    padding: 15px;
-    margin: 15px 0;
-    background-color: #fff;
-}
-
-h2 {
-    color: #007BFF;
-    margin-top: 0;
-}
-
-ul {
-    list-style-type: none;
-    padding-left: 0;
-}
-
-.in-stock {
-    color: green;
-}
-
-.out-of-stock {
-    color: red;
-}
