+++ src/pages/menuPage/MenuPage.css
@@ -1,0 +1,31 @@
+.item-list {
+    max-width: 600px;
+    margin: 0 auto;
+    padding: 20px;
+    font-family: Arial, sans-serif;
+    background-color: #f9f9f9;
+    border-radius: 6px;
+}
+
+h1 {
+    text-align: center;
+    color: #333;
+}
+
+.item {
+    border: 1px solid var(--color-blue);
+    border-radius: 6px;
+    padding: 15px;
+    margin: 15px 0;
+    background-color: var(--color-white);
+}
+
+h2 {
+    color: #08090A;
+    margin-top: 0;
+}
+
+ul {
+    list-style-type: none;
+    padding-left: 0;
+}
